Fix instruction memory for byte and half word accesses
Previously the LSBs of byte and half word accesses were ignore when the
access was not aligned with the start of a word. This commit also sets
initial values for `read_in_progress` and `write_in_progress` (which
were previously uninitialized on first read).

Note: the processor still does not support misaligned access.